interview analyzer/requirements.txt
flask==2.0.1
flask-cors==3.0.10
boto3==1.26.137
librosa==0.10.0.post2
numpy==1.23.5
moviepy==1.0.3
Pillow==9.5.0
transformers==4.28.1
torch==2.0.0
torchaudio==2.0.0
werkzeug==2.0.3
python-dotenv==0.19.0